cardgame/requirements.txt
asgiref==3.3.4
Django==3.2.4
pytz==2021.1
sqlparse==0.4.1
django-crispy-forms
django-widget-tweaks==1.4.8